+++ src/mlpack/core.hpp
@@ -254,6 +254,8 @@
  *   - Ayush Chamoli
  *   - Tommi Laivamaa <tommi.laivamaa@protonmail.com>
  *   - Kim SangYeon <sy0814k@gmail.com>
+ *   - Niteya Shah <niteya.56@gmail.com>
+ *   - Toshal Agrawal <tagrawal1339@gmail.com>
  */
 
 // First, include all of the prerequisites.