+++ src/mlpack/methods/ann/rbm/rbm.hpp
@@ -87,9 +87,10 @@ class RBM
    * parameters vector directly with Parameters() and modify it as desired.
    *
    * @param optimizer Optimizer type.
+   * @return The final objective of the trained model (NaN or Inf on error).
    */
   template<typename OptimizerType>
-  void Train(OptimizerType& optimizer);
+  double Train(OptimizerType& optimizer);
 
   /**
    * Evaluate the RBM network with the given parameters.