+++ src/mlpack/methods/nmf/nmf_main.cpp
@@ -28,10 +28,15 @@ using namespace mlpack::util;
 using namespace std;
 
 // Document program.
-PROGRAM_INFO("Non-negative Matrix Factorization", "This program performs "
-    "non-negative matrix factorization on the given dataset, storing the "
-    "resulting decomposed matrices in the specified files.  For an input "
-    "dataset V, NMF decomposes V into two matrices W and H such that "
+PROGRAM_INFO("Non-negative Matrix Factorization",
+    // Short description.
+    "An implementation of non-negative matrix factorization.  This can be used "
+    "to decompose an input dataset into two low-rank non-negative components.",
+    // Long description.
+    "This program performs non-negative matrix factorization on the given "
+    "dataset, storing the resulting decomposed matrices in the specified "
+    "files.  For an input dataset V, NMF decomposes V into two matrices W "
+    "and H such that "
     "\n\n"
     "V = W * H"
     "\n\n"
@@ -60,7 +65,17 @@ PROGRAM_INFO("Non-negative Matrix Factorization", "This program performs "
     PRINT_DATASET("H") + ", the following command could be used: "
     "\n\n" +
     PRINT_CALL("nmf", "input", "V", "w", "W", "h", "H", "rank", 10,
-        "update_rules", "multdist"));
+        "update_rules", "multdist"),
+    SEE_ALSO("@cf", "#cf"),
+    SEE_ALSO("Alternating matrix factorization tutorial",
+        "@doxygen/amftutorial.html"),
+    SEE_ALSO("Non-negative matrix factorization on Wikipedia",
+        "https://en.wikipedia.org/wiki/Non-negative_matrix_factorization"),
+    SEE_ALSO("Algorithms for non-negative matrix factorization (pdf)",
+        "http://papers.nips.cc/paper/1861-algorithms-for-non-negative-matrix-"
+        "factorization.pdf"),
+    SEE_ALSO("mlpack::amf::AMF C++ class documentation",
+        "@doxygen/classmlpack_1_1amf_1_1AMF.html"));
 
 // Parameters for program.
 PARAM_MATRIX_IN_REQ("input", "Input dataset to perform NMF on.", "i");
@@ -79,6 +94,41 @@ PARAM_STRING_IN("update_rules", "Update rules for each iteration; ( multdist | "
 
 PARAM_MATRIX_IN("initial_w", "Initial W matrix.", "p");
 PARAM_MATRIX_IN("initial_h", "Initial H matrix.", "q");
+
+void LoadInitialWH(const bool bindingTransposed, arma::mat& w, arma::mat& h)
+{
+  // Note that these datasets will typically be transposed on load, since we are
+  // likely receiving it from a row-major language, but we get it in a
+  // column-major form.  Therefore, we're actually decomposing V^T = W^T * H^T.
+  // Effectively this means we are solving, for the user, V = H*W.  Therefore,
+  // we actually have to switch what we are saving, so we will save the W we get
+  // from amf.Apply() as H, and vice versa.
+  if (bindingTransposed)
+  {
+    w = CLI::GetParam<arma::mat>("initial_h");
+    h = CLI::GetParam<arma::mat>("initial_w");
+  }
+  else
+  {
+    h = CLI::GetParam<arma::mat>("initial_h");
+    w = CLI::GetParam<arma::mat>("initial_w");
+  }
+}
+
+void SaveWH(const bool bindingTransposed, arma::mat&& w, arma::mat&& h)
+{
+  // The same transposition applies when saving.
+  if (bindingTransposed)
+  {
+    CLI::GetParam<arma::mat>("w") = std::move(h);
+    CLI::GetParam<arma::mat>("h") = std::move(w);
+  }
+  else
+  {
+    CLI::GetParam<arma::mat>("h") = std::move(h);
+    CLI::GetParam<arma::mat>("w") = std::move(w);
+  }
+}
 
 static void mlpackMain()
 {
@@ -105,7 +155,8 @@ static void mlpackMain()
   RequireAtLeastOnePassed({ "h", "w" }, false, "no output will be saved");
   RequireNoneOrAllPassed({"initial_w", "initial_h"}, true);
 
-  // Load input dataset.
+  // Load input dataset.  We know if the data is transposed based on the
+  // BINDING_MATRIX_TRANSPOSED macro, which will be 'true' or 'false'.
   arma::mat V = std::move(CLI::GetParam<arma::mat>("input"));
 
   arma::mat W;
@@ -121,9 +172,10 @@ static void mlpackMain()
     if (CLI::HasParam("initial_w"))
     {
       // Initialization with given W, H matrices.
-      GivenInitialization ginit = GivenInitialization(
-          std::move(CLI::GetParam<arma::mat>("initial_w")),
-          std::move(CLI::GetParam<arma::mat>("initial_h")));
+      arma::mat initialW, initialH;
+      LoadInitialWH(BINDING_MATRIX_TRANSPOSED, initialW, initialH);
+      GivenInitialization ginit = GivenInitialization(initialW, initialH);
+
       AMF<SimpleResidueTermination,
           GivenInitialization> amf(srt, ginit);
       amf.Apply(V, r, W, H);
@@ -143,9 +195,10 @@ static void mlpackMain()
     if (CLI::HasParam("initial_w"))
     {
       // Initialization with given W, H matrices.
-      GivenInitialization ginit = GivenInitialization(
-          std::move(CLI::GetParam<arma::mat>("initial_w")),
-          std::move(CLI::GetParam<arma::mat>("initial_h")));
+      arma::mat initialW, initialH;
+      LoadInitialWH(BINDING_MATRIX_TRANSPOSED, initialW, initialH);
+      GivenInitialization ginit = GivenInitialization(initialW, initialH);
+
       AMF<SimpleResidueTermination,
           GivenInitialization,
           NMFMultiplicativeDivergenceUpdate> amf(srt, ginit);
@@ -168,9 +221,10 @@ static void mlpackMain()
     if (CLI::HasParam("initial_w"))
     {
       // Initialization with given W, H matrices.
-      GivenInitialization ginit = GivenInitialization(
-          std::move(CLI::GetParam<arma::mat>("initial_w")),
-          std::move(CLI::GetParam<arma::mat>("initial_h")));
+      arma::mat initialW, initialH;
+      LoadInitialWH(BINDING_MATRIX_TRANSPOSED, initialW, initialH);
+      GivenInitialization ginit = GivenInitialization(initialW, initialH);
+
       AMF<SimpleResidueTermination,
           GivenInitialization,
           NMFALSUpdate> amf(srt, ginit);
@@ -185,9 +239,7 @@ static void mlpackMain()
     }
   }
 
-  // Save results.
-  if (CLI::HasParam("w"))
-    CLI::GetParam<arma::mat>("w") = std::move(W);
-  if (CLI::HasParam("h"))
-    CLI::GetParam<arma::mat>("h") = std::move(H);
+  // Save results.  Remember from our discussion in the comments earlier that we
+  // may need to switch the names of the outputs.
+  SaveWH(BINDING_MATRIX_TRANSPOSED, std::move(W), std::move(H));
 }